requirements.txt
torch==1.7.1+cu110
pytorch-lightning==1.0.5
pandas==1.1.4
pyarrow==2.0.0
scikit-learn==0.23.2
hydra-core==1.0.4
GitPython==3.1.11
plotly==4.13.0
matplotlib
